Paintbrush
Added tool panel inside of UXML which has a button to change to Paintbrush mode. Also tested a functionality where you can stop the user from using the paintbrush if they already have a specific prefab selected.

Added functionality to paintbrush which allows the function to be called when the button is pressed.

Improvements:

Need to set a somewhat mesh to follow the mouse cursor which can indicate where you are painting, could possibly use Unity's built in brushes as a template.

Will need to also allow the user to change what prefab they are painting during runtime.